Fix code bug in Text to avoid rasterization when drawing glyphs.

+++ intern/blendint/include/BlendInt/Gui/Context.hpp
@@ -267,12 +267,12 @@ namespace BlendInt {
 		static glm::mat4 default_view_matrix;
 	};
 
-	inline float pixel_size (int a)
+	inline int pixel_size (int a)
 	{
 		return a * Context::theme->pixel();
 	}
 
-	inline float pixel_size (unsigned int a)
+	inline int pixel_size (unsigned int a)
 	{
 		return a * Context::theme->pixel();
 	}
